+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Added: Option to include the script stack trace in the error log message (Write-ErrorLog)
+
 ## 3.3.0 - 2019-11-05
 
 * Added: The parent folder will be created, if it does not exists